Move core classes to xti_core
