+++ Lib/XTI_App.Api/ValueTemplateFromType.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using XTI_Core;
 
 namespace XTI_App.Api
 {
