+++ Lib/XTI_App/AppEventRepository.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using XTI_Core;
 
 namespace XTI_App
 {
