+++ Lib/XTI_App/AppRequest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using XTI_Core;
 
 namespace XTI_App
 {
