+++ Lib/XTI_App/AppSetup.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using XTI_Core;
 
 namespace XTI_App
 {
