+++ Lib/XTI_Core.Fakes/XTI_Core.Fakes.csproj
@@ -1,12 +1,8 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
   </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.8" />
-  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\XTI_Core\XTI_Core.csproj" />
