+++ Lib/XTI_Core/AppEventSeverity.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace XTI_App
+namespace XTI_Core
 {
     public sealed class AppEventSeverity : NumericValue, IEquatable<AppEventSeverity>
     {
