+++ Lib/XTI_Core/UtcClock.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace XTI_App
+namespace XTI_Core
 {
     public sealed class UtcClock : Clock
     {
