+++ Lib/XTI_WebApp.Extensions/XtiMiddleware.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using XTI_App;
 using XTI_App.Api;
+using XTI_Core;
 
 namespace XTI_WebApp.Extensions
 {
