+++ Projects/XTI_Version/ManageVersionCommand.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Text.Json;
 using System.Threading.Tasks;
 using XTI_App;
+using XTI_Core;
 
 namespace XTI_Version
 {
