+++ Tests/XTI_WebApp.AspTests/XtiMiddlewareTest.cs
@@ -19,6 +19,8 @@ using XTI_App;
 using XTI_App.Api;
 using XTI_App.EF;
 using XTI_Configuration.Extensions;
+using XTI_Core;
+using XTI_Core.Fakes;
 using XTI_WebApp.Extensions;
 using XTI_WebApp.Fakes;
 using XTI_WebApp.TestFakes;
