+++ Tests/XTI_WebApp.IntegrationTests/AppRoleIntegrationTest.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using XTI_App;
 using XTI_App.EF;
 using XTI_Configuration.Extensions;
+using XTI_Core;
 using XTI_WebApp.Extensions;
 using XTI_WebApp.Fakes;
