+++ Tests/XTI_WebApp.Tests/AuthorizationTest.cs
@@ -3,6 +3,7 @@ using NUnit.Framework;
 using System.Threading.Tasks;
 using XTI_App;
 using XTI_App.Api;
+using XTI_Core.Fakes;
 using XTI_WebApp.Fakes;
 
 namespace XTI_WebApp.Tests
