+++ Tests/XTI_WebApp.Tests/CachedUserContextTest.cs
@@ -13,6 +13,7 @@ using System.Threading.Tasks;
 using XTI_App;
 using XTI_App.Api;
 using XTI_App.EF;
+using XTI_Core;
 using XTI_WebApp.Extensions;
 using XTI_WebApp.Fakes;
 using XTI_WebApp.TestFakes;
